requirements.txt
# This should normally be a relatively short requirements.txt

# If you think of adding scikit-learn or tensorflow here, think again!
# This is the front-end. Heavy calculations go in the back-end, no?

# Sreamlit and extensions
streamlit
requests
plotly.express

# protobuf==3.19.0


# If you want to display datasets, or if your API returns you dataframes,
# you might need to add some extra stuff hereunder, e.g. pandas

# Data science
# pandas
